Clarify minimum size requirements for SD Card
I was not able to get Hassio to install / complete the installation process on an 8GB card.  I could not find any mention of minimum size requirements.  I cannot verify 16gb, but I assume that is the case.

I also think it would be helpful to mention how the process works.  It seems that a lot of people assume the HA splash screen will change during or after setup is complete.